Parameterized initialization
When initializing the build environment, we can disable the configuration of:
- MPI
- Doxygen

The reason is that the various projects all rely on CMake and LLVM, but not all
subprojects make use of MPI and Doxygen.
Since configuring all this stuff takes time (especially on Windows and Mac), we
can disable unneeded stuff to make tests run faster.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>